requirements.txt
flask
requests
Pillow
wordcloud